Move the global value numberer to the instruction selector - it makes more sense here
