+++ oc/compiler/optimizer/optimizer.c
@@ -7,7 +7,6 @@
 #include "optimizer.h"
 #include "../utils/queue/heap_queue.h"
 #include "../utils/constants.h"
-#include "../utils/value_numbering_table/value_numbering_table.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <sys/select.h>
@@ -2699,335 +2698,6 @@ static void optimize_short_circuit_logic(symtab_function_record_t* function, dyn
 
 
 /**
- * Get the value name for a given variable and *concatenate* it into
- * a given dynamic string. It is a assumed that the string has been
- * allocated by the caller
- */
-static inline void get_value_name(three_addr_var_t* variable, dynamic_string_t* output){
-	//Allocate a temporary buffer for this
-	char buffer[1000];
-	//Holder for the variable record
-	symtab_variable_record_t* variable_record = variable->linked_var;
-
-	//Handle each variable type accordingly
-	switch(variable->variable_type){
-		/**
-		 * Temporary variables just output as t<number>
-		 */
-		case VARIABLE_TYPE_TEMP:
-			sprintf(buffer, "t%d", variable->temp_var_number);
-			dynamic_string_concatenate(output, buffer);
-
-			break;
-
-		/**
-		 * For non temporaries we will use:
-		 * 	<lexical_scope>_name_<ssa_generation>
-		 *
-		 * 	This will guarantee uniqueness even if we have
-		 * 	colliding
-		 */
-		case VARIABLE_TYPE_NON_TEMP:
-			//Store the variable record
-			sprintf(buffer, "%d_%s_%d", variable_record->lexical_scope_id, variable_record->var_name.string, variable->ssa_generation);
-			dynamic_string_concatenate(output, buffer);
-
-			break;
-
-		/**
-		 * For a memory address, we will just print this
-		 * out as M<<lexical_scope>_<name>_<ssa_generation>>
-		 * if we have a variable name. If not then we'll just be printing
-		 * out the temp var number
-		 */
-		case VARIABLE_TYPE_MEMORY_ADDRESS:
-			if(variable_record != NULL){
-				sprintf(buffer, "M<%d_%s_%d>", variable_record->lexical_scope_id, variable_record->var_name.string, variable->ssa_generation);
-			} else {
-				sprintf(buffer, "M<t%d>", variable->temp_var_number);
-			}
-
-			dynamic_string_concatenate(output, buffer);
-			break;
-
-		/**
-		 * For a stack memory address, we will just print this
-		 * out as SM<<scope>_<name>_<ssa_generation>> if we have a
-		 * variable name. If not then we will be using the temp 
-		 * variable number
-		 */
-		case VARIABLE_TYPE_STACK_PARAM_MEMORY_ADDRESS:
-			if(variable_record != NULL){
-				sprintf(buffer, "SM<%d_%s_%d>", variable_record->lexical_scope_id, variable_record->var_name.string, variable->ssa_generation);
-			} else {
-				sprintf(buffer, "SM<t%d>", variable->temp_var_number);
-			}
-
-			dynamic_string_concatenate(output, buffer);
-			break;
-
-		/**
-		 * Local constants are already unique so we can just print them
-		 * out as is
-		 */
-		case VARIABLE_TYPE_LOCAL_CONSTANT:
-			sprintf(buffer, ".LC%d", variable->associated_memory_region.local_constant->local_constant_id);
-			dynamic_string_concatenate(output, buffer);
-
-			break;
-
-		/**
-		 * Function addresses just output the name of the underlying function
-		 */
-		case VARIABLE_TYPE_FUNCTION_ADDRESS:
-			sprintf(buffer, "%s", variable->associated_memory_region.rip_relative_function->func_name.string);
-			dynamic_string_concatenate(output, buffer);
-			break;
-
-		default:
-			fprintf(stderr, "Fatal internal compiler error: unrecognized variable type detected in value name generator\n");
-			exit(1);
-	}
-}
-
-/**
- * Is the given phi function redundant? A phi function is redundant
- * if *all* of the variables inside of the phi function have ended
- * up becoming the same via the GVN pass. This is easy to check
- * because we just need to compare the phi function's first parameter
- * against all of the others. If they all match, then it is redundant
- * and we will convert it into an assignment
- *
- * Return TRUE if it was redundant, FALSE if not. If it was redundant then
- * we will have transformed this into an assignment
- */
-static inline u_int8_t convert_phi_function_if_redundant(instruction_t* phi_function){
-	//Extract the parameters
-	dynamic_array_t* parameters = &(phi_function->parameters);
-
-	//Grab the first parameter
-	three_addr_var_t* first_parameter = dynamic_array_get_at(parameters, 0);
-
-	//Now run through every other parameter and compare
-	for(u_int32_t i = 1; i < parameters->current_index; i++){
-		three_addr_var_t* compare_to_param = dynamic_array_get_at(parameters, i);
-
-		//If these aren't equal then we fail out
-		if(variables_equal(first_parameter, compare_to_param, FALSE) == FALSE){
-			return FALSE;
-		}
-	}
-
-	/**
-	 * If we survived to down here then the phi function is redundant. We will replace
-	 * this phi function with a regular copy assignment now
-	 */
-	phi_function->statement_type = THREE_ADDR_CODE_ASSN_STMT;
-	phi_function->op1 = first_parameter;
-
-	//Deallocate the parameters array
-	dynamic_array_dealloc(&(phi_function->parameters));
-
-	//This was redundant
-	return TRUE;
-}
-
-
-/**
- * Generate the key for a given instruction. The key for a given
- * instruction consists of the instruction type converted to a char *and*
- * the variables that are inside of it. We are given an allocated
- * dynamic string as the key for this function and we will populate it
- */
-static inline void generate_value_name_key_for_instruction(instruction_t* instruction, dynamic_string_t* textual_key){
-	//Based on the instruction type we generate different keys
-	switch(instruction->instruction_type){
-		case THREE_ADDR_CODE_PHI_FUNC:
-			dynamic_string_concatenate(textual_key, "PHI");
-
-			//Concatenate the variable name of each of the parameters onto the end
-			for(u_int32_t i = 0; i < instruction->parameters.current_index; i++){
-				get_value_name(dynamic_array_get_at(&(instruction->parameters), i), textual_key);
-			}
-
-			break;
-
-		//TODO NOT IN YET
-		default:
-			break;
-	}
-}
-
-
-/**
- * Perform the value numbering algorithm for one block. This algorithm
- * naturally reproduces itself in a recursive manner by traversing the dominator tree.
- * Traversing a dominator tree will naturally traverse the function in reverse post
- * order
- *
- * NOTE: this function is recursive
- *
- * Algorithm Dominator Value Numbering Traversal(block b):
- * 	for each phi node in b:
- * 		if phi node is redundant then:
- * 			remove it
- * 			continue
- *
- * 		set the value number as the assigned var name
- * 		add phi node to hash table
- *
- * 	for each assignment i:
- * 		get value numbers for each operand
- * 		if the expression(Ti) has been computed before:
- * 			replace the operation i with a copy from Ti
- * 			associate the value number with Ti
- * 		else:
- * 			insert a new value number into the table at the hash key location
- * 			record that new value number for Ti
- *
- * 	for each successor c of block b:
- * 		replace all phi node operands in c that were computed in this block with their value
- *
- * 	for each child c of block b in the dominator tree
- * 		Dominator Value Numbering Traversal(b)
- */
-static void global_value_number_block(value_numbering_table_t* table, basic_block_t* block){
-	/**
-	 * 	for each phi node in b:
-	 * 		if phi node is redundant then:
-	 * 			convert it to an assignment
-	 * 			continue
-	 *
-	 * 		set the value number as the assigned var name
-	 * 		add phi instruction to hash table
-	 *
-	 * 	For a phi node to be redundant, all of the values inside of
-	 * 	it have to be the exact same
-	 */
-	instruction_t* cursor = block->leader_statement;
-
-	//So long as we see phi functions
-	while(cursor != NULL && cursor->statement_type == THREE_ADDR_CODE_PHI_FUNC){
-		//It's redundant so we continue out
-		if(convert_phi_function_if_redundant(cursor) == TRUE){
-			cursor = cursor->next_statement;
-			continue;
-		}
-		
-		/**
-		 * Otherwise we will need to save this value inside of the hash table. We
-		 * will generate the value name for it and save it inside. We need to do
-		 * this so that we have a trail of where it came from
-		 */
-
-
-		//Bump it up
-		cursor = cursor->next_statement;
-	}
-
-
-
-	/**
-	 * For each child c of the block in the *dominator* tree, we will
-	 * invoke this same algorithm recursively
-	 */
-	for(u_int32_t i = 0; i < block->dominator_children.current_index; i++){
-		//Extract the dominator chid
-		basic_block_t* dominator_child = dynamic_array_get_at(&(block->dominator_children), i);
-
-		//Recursively explore this one next
-		global_value_number_block(table, dominator_child);
-	}
-}
-
-
-/**
- * Estimate the keyspace that we'll need for a given function
- * by tallying up the number of instructions that the function contains.
- * This is a very rough estimate, it's just designed to help us
- * avoid allocating tons of space
- */
-static inline u_int32_t estimate_value_numbering_keyspace_for_function(dynamic_array_t* function_blocks){
-	//Initially we have nothing
-	u_int32_t keyspace = 0;
-
-	//Run through every block in the function
-	for(u_int32_t i = 0; i < function_blocks->current_index; i++){
-		basic_block_t* function_block = dynamic_array_get_at(function_blocks, i);
-
-		//Tally up the total instructions
-		keyspace += function_block->number_of_instructions;
-	}
-
-	/**
-	 * If we find that we have less than a certain
-	 * number of instructions, we'll do nothing because
-	 * we expect to not be value numbering here. However
-	 * if we exceed the threshold, we will double the
-	 * keyspace to try and cut done on collisions
-	 */
-	if(keyspace <= INSTRUCTION_NUMBER_THRESHOLD){
-		return keyspace;
-	} else {
-		return keyspace * 2;
-	}
-}
-
-
-/**
- * Perform a global value numbering pass to determine if there are any redundant computations. This relies on
- * everything being in SSA form which OIR uses by default. We will also need the dominator tree and the ability
- * to traverse in reverse post order. We will be using the SSA names as the value numbers themselves inside of our
- * contextual strings
- *
- * Algorithm Dominator Value Numbering Traversal(block b):
- * 	for each phi node in b:
- * 		set the value number as the assigned var name
- * 		add phi node to hash table
- *
- * 	for each assignment i:
- * 		get value numbers for each operand
- * 		if the expression(Ti) has been computed before:
- * 			replace the operation i with a copy from Ti
- * 			associate the value number with Ti
- * 		else:
- * 			insert a new value number into the table at the hash key location
- * 			record that new value number for Ti
- *
- * 	for each successor c of block b:
- * 		replace all phi node operands in c that were computed in this block with their value
- *
- * 	for each child c of block b in the dominator tree
- * 		Dominator Value Numbering Traversal(b)
- *
- * We will be using the full SSA variable names as their given value numbers inside of this algorithm. For temporary
- * variables, their temp var number(which is guaranteed unique across the entire program) will be used as the value
- * number. This should avoid any/all collisions that we may face
- *
- * For right now, we will limit the value numbering to non-constant operations. In the future we will probably
- * expand this to include constants as well
- */
-static void global_value_numbering_pass(symtab_function_record_t* function, basic_block_t* function_entry_block, dynamic_array_t* function_blocks){
-	//Estimate the keyspace first
-	u_int32_t keyspace = estimate_value_numbering_keyspace_for_function(function_blocks);
-
-	//If we are below the threshold, we'll skip this function as it almost certainly isn't worth it
-	if(keyspace <= INSTRUCTION_NUMBER_THRESHOLD){
-		return;
-	}
-
-	//Otherwise we can allocate our hash table
-	value_numbering_table_t numbering_table = value_numbering_table_alloc(keyspace);
-	
-	//Invoke the value numberer with the function entry block as our starting point
-	global_value_number_block(&numbering_table, function_entry_block);
-
-	//Destroy the table once down
-	value_numbering_table_dealloc(&numbering_table);
-}
-
-
-/**
  * Is a given conditional always true or always false? We will need
  * to trace up the block to find out. If we are unable to
  * find out, that is ok, we just return false and assume
@@ -3505,14 +3175,6 @@ cfg_t* optimize(cfg_t* cfg){
 		 * that we can expend the compute of optimizing the short circuit logic on them, and we will do so here
 		 */
 		optimize_short_circuit_logic(current_function, current_function_blocks);
-
-		/**
-		 * PASS XX: function global value numbering pass
-		 * Now that we've marked and swept everything, we will use global value numbering within the function to
-		 * remove any redundant calculations. This is done after mark and sweep because we don't want
-		 * to be doing this operations for values that end up being useless anyways
-		 */
-		global_value_numbering_pass(current_function, function_entry_block, current_function_blocks);
 
 		/**
 		 * PASS 4: always true/false optimization
